Addresses feedback
