+++ app/controllers/followers_controller.rb
@@ -46,18 +46,6 @@ class FollowersController < ApplicationController
 
   def create
     follower = create_follower(params)
-    if request.format.json? && !request.headers["X-Inertia"]
-      return render json: { success: false, message: "Sorry, something went wrong." } if follower.nil?
-      return render json: { success: false, message: follower.errors.full_messages.to_sentence } if follower.errors.present?
-
-      if follower.confirmed?
-        render json: { success: true, message: "You are now following #{follower.user.name_or_username}!" }
-      else
-        render json: { success: true, message: "Check your inbox to confirm your follow request." }
-      end
-      return
-    end
-
     followed_user = User.find_by_external_id(params[:seller_id])
     fallback_url = followed_user&.profile_url || root_path
