+++ app/javascript/pages/Users/Show.tsx
@@ -1,25 +1,10 @@
 import { usePage } from "@inertiajs/react";
 import * as React from "react";
 
-import { Profile, Props as ProfileProps } from "$app/components/server-components/Profile";
-import BasePage from "$app/utils/base_page";
-
-type Props = ProfileProps & {
-  card_data_handling_mode: string;
-  paypal_merchant_currency: string;
-};
-
-let initialized = false;
+import { Profile, Props as ProfileProps } from "$app/components/Profile";
 
 export default function UsersShowPage() {
-  const props = usePage<Props>().props;
-
-  React.useEffect(() => {
-    if (!initialized) {
-      BasePage.initialize();
-      initialized = true;
-    }
-  }, []);
+  const props = usePage<ProfileProps>().props;
 
   return (
     <div className="flex h-screen flex-col overflow-y-auto">
