+++ app/javascript/ssr.ts
@@ -17,13 +17,6 @@ import Nav from "$app/components/server-components/Nav";
 import ProductPage from "$app/components/server-components/Product";
 import ProductIframePage from "$app/components/server-components/Product/IframePage";
 import ProductEditPage from "$app/components/server-components/ProductEditPage";
-import ProfileCoffeePage from "$app/components/server-components/Profile/CoffeePage";
-import ProfileProductPage from "$app/components/server-components/Profile/ProductPage";
-import ProfileWishlistPage from "$app/components/server-components/Profile/WishlistPage";
-import DisputeEvidencePage from "$app/components/server-components/Purchase/DisputeEvidencePage";
-import PurchaseProductPage from "$app/components/server-components/Purchase/ProductPage";
-import SecureRedirectPage from "$app/components/server-components/SecureRedirectPage";
-import SubscribePreview from "$app/components/server-components/SubscribePreview";
 import SubscriptionManager from "$app/components/server-components/SubscriptionManager";
 import SubscriptionManagerMagicLink from "$app/components/server-components/SubscriptionManagerMagicLink";
 import SupportHeader from "$app/components/server-components/support/Header";
@@ -50,11 +43,6 @@ ReactOnRails.register({
   ProductEditPage,
   ProductIframePage,
   ProductPage,
-  ProfileCoffeePage,
-  ProfileProductPage,
-  ProfileWishlistPage,
-  SecureRedirectPage,
-  SubscribePreview,
   SubscriptionManager,
   SubscriptionManagerMagicLink,
   TaxesCollectionModal,
